requirements.txt
python-telegram-bot
yt-dlp
requests
beautifulsoup4
python-dotenv

